requirements.txt
# Core
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.5.2
httpx==0.27.2

# Data / utils
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.5

# LLM providers
openai==1.61.1
anthropic==0.45.2
google-genai==1.0.0

# App / demo
streamlit==1.41.1

# Utilities
rich==13.9.4
tenacity==9.0.0
pypdf==4.3.1
markdown==3.7
beautifulsoup4==4.12.3

# API Integration
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Vector store
faiss-cpu==1.8.0

# Testing
pytest==8.3.3
pytest-cov==5.0.0